Specialize for windows
